Add some Cabinet items; KeyCode/Position refactoring

+++ core/src/main/scala/org/somecode/enig4s/mach/WheelState.scala
@@ -5,4 +5,4 @@ final case class WheelState(
   position: KeyCode,
   ring: RingSetting
 ):
-  val offset: Int = position - ring
+  val offset: Int = position.toInt - ring.toInt
